third_party/WebKit/LayoutTests/external/wpt/beacon/headers/header-content-type-expected.txt

This is a testharness.js-based test.
PASS Test content-type header for a body string
FAIL Test content-type header for a body ArrayBufferView assert_equals: Correct Content-Type header result expected "" but got "application/octet-stream"
FAIL Test content-type header for a body ArrayBuffer assert_equals: Correct Content-Type header result expected "" but got "text/plain;charset=UTF-8"
PASS Test content-type header for a body Blob
PASS Test content-type header for a body FormData
FAIL Test content-type header for a body URLSearchParams assert_equals: Correct Content-Type header result expected "application/x-www-form-urlencoded;charset=UTF-8" but got "text/plain;charset=UTF-8"
Harness: the test ran to completion.
